# openapi/newton-backend-parity.yaml
openapi: 3.1.0
info:
  title: Newton Backend Parity API
  description: Contract for Rust `newton serve` parity endpoints sourced from Rust DTOs and handlers.
  contact:
    name: Newton Loop Framework
  license:
    name: ''
  version: 0.5.101
servers:
- url: /api/v1
paths:
...
  /executions:
    get:
      tags:
      - executions
      operationId: list_executions
      parameters:
      - name: planId
        in: query
        description: Optional plan id filter
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Execution list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExecutionItem'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /opportunities:
    get:
      tags:
      - opportunities
      operationId: list_opportunities
      parameters:
      - name: status
        in: query
        description: Optional opportunity status filter
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Opportunity list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OpportunityItem'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /opportunities/{id}:
    patch:
      tags:
      - opportunities
      operationId: patch_opportunity
      parameters:
      - name: id
        in: path
        description: Opportunity id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchOpportunityBody'
        required: true
      responses:
        '200':
          description: Updated opportunity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityItem'
        '404':
          description: Opportunity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /persistence/{key}:
    get:
      tags:
      - persistence
      operationId: get_persistence
      parameters:
      - name: key
        in: path
        description: Persistence key
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Persisted JSON value
          content:
            application/json:
              schema: {}
        '404':
          description: Key not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    put:
      tags:
      - persistence
      operationId: put_persistence
      parameters:
      - name: key
        in: path
        description: Persistence key
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema: {}
        required: true
      responses:
        '200':
          description: Persistence upsert result
          content:
            application/json:
              schema: {}
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
      - persistence
      operationId: delete_persistence
      parameters:
      - name: key
        in: path
        description: Persistence key
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Persistence delete result
          content:
            application/json:
              schema: {}
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /plans:
    get:
      tags:
      - plans
      operationId: list_plans
      responses:
        '200':
          description: Plan list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlanItem'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /plans/{id}:
    get:
      tags:
      - plans
      operationId: get_plan
      parameters:
      - name: id
        in: path
        description: Plan id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Plan detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanDetail'
        '404':
          description: Plan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /plans/{id}/approve:
    post:
      tags:
      - plans
      operationId: approve_plan
      parameters:
      - name: id
        in: path
        description: Plan id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Approved plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanItem'
        '404':
          description: Plan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Plan state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /plans/{id}/reject:
    post:
      tags:
      - plans
      operationId: reject_plan
      parameters:
      - name: id
        in: path
        description: Plan id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Rejected plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanItem'
        '404':
          description: Plan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Plan state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /testing/reset:
    post:
      tags:
      - testing
      operationId: reset_testing
      responses:
        '200':
          description: Reset result
          content:
            application/json:
              schema: {}
        '403':
          description: Forbidden in production
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
    ApiError:
      type: object
      required:
      - code
      - category
      - message
      properties:
        category:
          type: string
        code:
          type: string
        details: {}
        message:
          type: string
...
    ExecutionItem:
      type: object
      required:
      - instance_id
      - status
      - created_at
      properties:
        component:
          type:
          - string
          - 'null'
        created_at:
          type: string
        deployStatus:
          type:
          - string
          - 'null'
        instance_id:
          type: string
        linkedPlanId:
          type:
          - string
          - 'null'
        planId:
          type:
          - string
          - 'null'
        planTitle:
          type:
          - string
          - 'null'
        policyLevel:
          type:
          - string
          - 'null'
        prLink:
          type:
          - string
          - 'null'
        prStatus:
          type:
          - string
          - 'null'
        repo:
          type:
          - string
          - 'null'
        stage:
          type:
          - string
          - 'null'
        started:
          type:
          - string
          - 'null'
        startedBy:
          type:
          - string
          - 'null'
        status:
          type: string
        testResult:
          type:
          - string
          - 'null'
        waitingOn:
          type:
          - string
          - 'null'
        workflow_id:
          type:
          - string
          - 'null'
...
    OpportunityItem:
      type: object
      required:
      - id
      - title
      - origin
      - component
      - repo
      - risk
      - expectedValue
      - status
      - dependsOn
      - blocks
      properties:
        age:
          type:
          - string
          - 'null'
        blocks:
          type: array
          items:
            type: string
        component:
          type: string
        confidence:
          type:
          - number
          - 'null'
          format: double
        dependsOn:
          type: array
          items:
            type: string
        effort:
          type:
          - string
          - 'null'
        expectedValue:
          type: number
          format: double
        id:
          type: string
        indicator:
          type:
          - string
          - 'null'
        module:
          type:
          - string
          - 'null'
        origin:
          type: string
        rationale:
          type:
          - string
          - 'null'
        repo:
          type: string
        risk:
          type: string
        status:
          type: string
        title:
          type: string
...
    PatchOpportunityBody:
      type: object
      required:
      - status
      properties:
        status:
          type: string
...
    PlanApproverItem:
      type: object
      required:
      - role
      - name
      - status
      properties:
        name:
          type: string
        role:
          type: string
        status:
          type: string
    PlanDetail:
      allOf:
      - $ref: '#/components/schemas/PlanItem'
      - type: object
        required:
        - sections
        - policyChecks
        - approvers
        properties:
          approvers:
            type: array
            items:
              $ref: '#/components/schemas/PlanApproverItem'
          expectedDelta:
            type:
            - string
            - 'null'
          policyChecks:
            type: array
            items:
              $ref: '#/components/schemas/PlanPolicyCheckItem'
          sections:
            type: array
            items:
              $ref: '#/components/schemas/PlanSectionItem'
    PlanItem:
      type: object
      required:
      - id
      - title
      - component
      - repo
      - status
      - executionIds
      - confidence
      - risk
      - agentGenerated
      - createdAt
      properties:
        agentGenerated:
          type: boolean
        component:
          type: string
        confidence:
          type: integer
          format: int64
        createdAt:
          type: string
        executionIds:
          type: array
          items:
            type: string
        expectedValue:
          type:
          - string
          - 'null'
        id:
          type: string
        linkedRequestId:
          type:
          - string
          - 'null'
        repo:
          type: string
        risk:
          type: string
        status:
          type: string
        title:
          type: string
        waitingSince:
          type:
          - string
          - 'null'
    PlanPolicyCheckItem:
      type: object
      required:
      - rule
      - status
      - met
      properties:
        met:
          type: boolean
        rule:
          type: string
        status:
          type: string
    PlanSectionItem:
      type: object
      required:
      - id
      - label
      - content
      properties:
        content:
          type: string
        id:
          type: string
        label:
          type: string
...